Update README.md
Snake Xenia is a game we used to play in Mobiles. I have tried to implement the game on console by using MASM and Assembely Language following Irvine32 library. I am a Computer Science student and as we know that we all have to write program to understand the better about the field and to create programmes essential for of our daily routines. I am a CS student at National University Of Science and Technology (NUCES) , Karachi Campus ,Pakistan.